+++ src/adb.spec.js
@@ -1,7 +1,8 @@
 "use strict";
 
 /*
- * Copyright (C) 2017-2021 UBports Foundation <info@ubports.com>
+ * Copyright (C) 2017-2022 UBports Foundation <info@ubports.com>
+ * Copyright (C) 2017-2022 Johannah Sprinz <hannah@ubports.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -353,7 +354,7 @@ describe("Adb module", function () {
           expect(child_process.spawn).toHaveBeenCalledWith(
             adb.executable,
             [...adb.extra, "sideload", "tests/test-data/test_file"],
-            { env: { ADB_TRACE: "rwx" } }
+            { env: expect.objectContaining({ ADB_TRACE: "rwx" }) }
           );
         });
       });
@@ -1017,6 +1018,7 @@ describe("Adb module", function () {
     it("should restore full backup", function () {
       stubExec(1, "should not be called");
       jest.useFakeTimers();
+      jest.setSystemTime();
       fs.readJSON = jest.fn().mockReturnValue({
         codename: "codename",
         comment: "Ubuntu Touch backup created on 1970-01-01T00:00:00.000Z",